requirements.txt
Django==1.10.3
djangorestframework==3.5.2
html5lib==0.999999999
ipython==5.1.0
beautifulsoup4==4.5.1
uuid==1.30
requests==2.11.1
lxml==3.6.4
